ui progress. added event system to models
