+++ UNFI File Search/settings.ini
@@ -1,11 +1,11 @@
 [DEFAULT]
-username = ""
-password = ""
+username = 
+password = 
 auto_download = False
 auto_login = True
-auto_save = True
+auto_save = False
 default_save_path = F:/POS/unfi/query_new.xlsx
 
 [SEARCH]
-search_chunk_size = 120
+search_chunk_size = 1737
