+++ UNFI File Search/unfi_gui/controller.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING, Dict, List, Union
 import tkinter as tk
 import tkinter.ttk as ttk
-
+from unfi_api.utils.threading import stop_job, stop_all_jobs
 
 if TYPE_CHECKING:
     from model import TkModel
@@ -17,7 +17,11 @@ class Controller:
         self.home_frame: str = None
         self.models: dict[str, TkModel] = {}
         self.ready: bool = False
+        self.jobs = []
+        self.cancel = []
+        self.cancel_all = False
 
+    
     def register_views(self, frames: List["View"]) -> None:
         self.container.setup(frames)
         self.setup = True
@@ -74,3 +78,27 @@ class Controller:
     
     def setup(self) -> None:
         ...
+    
+    def add_job_id(self, job_id: Union[str, int]) -> None:
+        self.jobs.append(job_id)
+    
+    def stop_job(self, job_id: Union[str, int]) -> None:
+        if job_id in self.jobs:
+            stop_job(job_id)
+
+    def stop_all_jobs(self) -> None:
+        for job_id in self.jobs:
+            stop_job(job_id)
+    
+    def stop_cancelled_jobs(self) -> None:
+        for job_id in self.cancel:
+            stop_job(job_id)
+            self.cancel.remove(job_id)
+    
+    def cancel_job(self, job_id: Union[str, int]) -> None:
+        self.cancel.append(job_id)
+    
+    def cancel_all_jobs(self) -> None:
+        self.cancel_all = True
+        self.stop_all_jobs()
+
