+++ settings.ini
@@ -4,7 +4,7 @@ password = ""
 auto_download = False
 auto_login = False
 auto_save = True
-default_save_path = "F:\\pos\\unfi\\query_new.xlsx"
+default_save_path = F:\\pos\\unfi\\query_new.xlsx
 
 [SEARCH]
 search_chunk_size = 120
